WIP http client

+++ include/peripheral/dma_serial.h
@@ -25,7 +25,7 @@ typedef struct {
 
 uint16_t dma_serial_count(DMA_SerialHandle* handle);
 uint8_t dma_serial_getc(DMA_SerialHandle* handle);
-uint8_t dma_serial_read(DMA_SerialHandle* handle, uint8_t* buffer, uint16_t position, uint16_t length);
+uint16_t dma_serial_read(DMA_SerialHandle* handle, uint8_t* buffer, uint16_t position, uint16_t length);
 
 #ifdef __cplusplus
 }
